Add files via upload
A queue is a container data structure that acts like a pipeline. That is, data enters one end of the pipe and is removed from the other end (called FIFO: first in, first out). A stack is similar to a queue except the data is removed (pop) from the same end it is inserted into (push) (called LIFO: last in, first out). Data cannot be removed from an empty queue nor from an empty stack. Also links which are removed from the list should be deleted from memory.
Build a menu-controlled application using various classes and linked lists (stacks and queues) and string data to 
1) add data to the input end of the queue, 
2) remove data from the output end of the queue 
   (also print each value as it is removed),
3) clear a queue of its contents, 
4) print the contents of the queue, 
5) reverse the data inside the queue (create a temporary linked stack, remove the data from   the queue and place it into the stack, then remove it from the stack and place it back into the queue). (There are other ways to do this…do it this way for practice.)
6) exit from the application

The GUI that you design is totally your decision.