Add files via upload
Application 1

Write a demonstration program (e.g. radiobutton-driven application) that can be used to call three functions that are solved using recursion.
Your program should provide a user interface (similar to the one below) from which the user can select to do any of the operations (by entering data in the textfield(s) and pressing the correct radio button). After your program does the requested operation, it should inform the user of the result in the text area and await the next request. 
Your application incorporates the use of (at least) the following classes: JFrame, JRadioButton , JTextField, JTextArea, JLabel, ButtonGroup, and Container.
In addition, make the text area uneditable. Tips to the user should appear each time the mouse pointer is placed over a radio button.
Function 1: Write a recursive method power (base, exponent) that when invoked returns base exponent.

For example, power(3,4) = 3 * 3 * 3 * 3. Assume that the exponent is an integer greater than or equal to 1. 



Application 2

Write an application including a recursive function that determines whether an 
array is palindrome, where the array and its size are given as parameters. 


	//returns 1 if a[ ] is a palindrome, 0 otherwise
	int ispalindrome(char a[ ], int n)


Application 3

Write an application including a recursive function that reverses the digits of any positive integer (not print the individual digits on the screen in reverse).

Please have the following available for collection during your practical session: